Implemented testing for IngredientServiceImpl

+++ Schollmeyer_Dan_ClayBrewing_CaseStudy/src/main/java/com/optum/Ingredient/IngredientRepo.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface IngredientRepo extends JpaRepository<Ingredient, Long> {
     Ingredient findFirstByName(String name);
-    Ingredient findFirstById(int id);
+    Ingredient findFirstById(long id);
 }
